Normalize the serialization and deserialization places of protobuf structs
